requirements.txt
boto3
requests
mysql-connector-python